+++ CHANGELOG.md
@@ -6,6 +6,7 @@ LaplacianOpt.jl Change Log
 - Fixed Docs compile issue
 - Updated 100 and 60 nodes instances with solutions
 - Minor updates in graphviz logging in `log.jl`
+- `plot_solutions.jl` script added to plot graphs
 
 ### v0.6.0
 - Added mutiple heuristics to handle both spanning trees and graphs with loops